Move useStripProductsFromUrl to own file
